services/nest/requirements.txt
framework==0.2.7
pymongo
azure-servicebus
swagger-gen
quart
uvloop
uvicorn
python-dotenv
aioredis
deprecated
httpx
motor